requirements.txt
requests
urllib3
websockets
wrapt
yarl
bs4
selenium
webdriver-manager
mysql-connector-python
matplotlib
nltk
pillow
validators
table2ascii
pyvirtualdisplay
undetected-chromedriver
flask
flask-limiter